package/ci: add a static build on Linux, macOS and Windows.
This was originally in a PR from 2020, where I attempted to make things
work with static libraries and dynamic plugins and (obviously) failed
spectacularly. Such a use case is rather rare due to its cursedness, and
it doesn't make sense to not have a static CI build at all just because
that use case is broken.

So, instead, let's just currently assume that if MAGNUM_BUILD_STATIC is
set, the plugins are static as well, and treat them as such. The Ui
library tests in magnum-extras do the same and it's fine.