EigenIntegration: split dynamic matrix conversion to a separate header.
Originally it was put into the same header as everything else because
including StridedArrayView in addition to everything from Eigen didn't
make the already-horrific compile times any worse. But for the
MagnumMath single-header library I want to be able to opt-out from
StridedArrayView as it would unconditionally add quite a lot of extra
dependencies to the file.

If MAGNUM_BUILD_DEPRECATED is enabled, Integration.h still includes the
new header for backward compatibility.