ImGuiIntegration: try to make this work with vcpkg again.
Initially, vcpkg was building imgui as a DLL (with symbols correctly
exported, and everything correct) in the default shared build. But
people had their projects broken because the thing didn't define
IMGUI_API correctly, so ultimately this got "fixed" as building imgui
statically always. Which then broke Magnum, because Magnum expects that
on a dynamic build, ImGui is also dynamic (because that's what makes
sense for not having duplicated globals, right).

Now FindImGui tries to query whether the imgui found via a config file
is static, and then defines IMGUI_API to empty to not have it assumed to
be dynamic if Magnum is dynamic.