ImGuiIntegration: use OpenGL texture ID inside ImTextureID, not a pointer.
A pointer was used since the very first contributed implementation, but a
misuse of ImTextureID or doing anything even slightly wrong when setting
up fonts led to a rather nasty crash where it attempted to call bind() on
a pointer that was definitely not a GL::Texture2D instance.

So instead it's an OpenGL ID now, which in the worst case gives a GL
error and renders nothing, but doesn't kill the whole app. Internally a
temporary non-owning GL::Texture2D instance is created when binding it to
the shader, in order to make state tracker happy and all that.

No change is needed for code that uses the image() / imageButton()
helpers as the change is done transparently inside, but code that passes
ImTextureID to low-level ImGui drawing APIs needs to be adapted. For
that, and for making it possible to do additional changes in the future
like supporting texture arrays, there's a new textureId() helper that you
should use instead of a manual cast.