CMake: rename all MAGNUM_WITH_* options to MAGNUM_WITH_*_INTEGRATION.
While, say, specifying MAGNUM_WITH_BULLET kind of makes sense when
building magnum-integration alone (as all libraries have an Integration
suffix here), it becomes confusing when building this as a subproject
along with other Magnum libraries, making it look like there's a
Magnum::Bullet library, or maybe that Magnum has a Bullet "backend" or
some such. Neither of that is the case, and given that in all other
cases the CMake option name matches the namespace, class or library
name, this is only a pointless special-case that hurts intuitive usage.

Case in point, in the CPM building docs, MAGNUM_WITH_IMGUIINTEGRATION
was already (accidentally) used as an example of where to specify such
options, meaning that not even I can remember that it's different here.

Unfortunately I realized this too late after doing the MAGNUM_* prefix
rename, so this needs a second level of backwards compatibility, dealing
with both WITH_BLAH -> MAGNUM_WITH_BLAHINTEGRATION and MAGNUM_WITH_BLAH
-> MAGNUM_WITH_BLAHINTEGRATION. Fortunately the original backwards
compat code is quite proven in practice already, so all I had to do was
to copy it for this new case.